Added Winning Logic - Column Scan
Scan through the column at the marked space.
I call the scanning space "tray"

Checks whether all the marks at the column
are same as the mark of the current player.
Returns Boolean

Added Winning Logic - Diagonal Scan
Scan through the diagonal from
Left  -> Right
and
Right -> Left

Checks whether all the marks at the both diagonal
are same as the mark of the current player.
Returns Boolean

Now changed the dimension from hardcoded 3x3 to
size x size. The default is 3x3

Added a Helper Method - all_same?

Checks whether all the elements of "tray"
are 0 or not. Returns boolean

Checks if all the elements of a